+++ core/src/main/kotlin/io/github/serpro69/kfaker/dictionary/Dictionary.kt
@@ -25,6 +25,7 @@ internal enum class CategoryName {
     ARTIST,
     BACK_TO_THE_FUTURE,
     BANK,
+    BARCODE,
     BASKETBALL,
     BEER,
     BLOOD,