+++ core/src/main/kotlin/io/github/serpro69/kfaker/Faker.kt
@@ -65,6 +65,7 @@ class Faker @JvmOverloads constructor(internal val fakerConfig: FakerConfig = Fa
     val dog: Dog
     val dota: Dota
     val dragonBall: DragonBall
+    val drivingLicense: DrivingLicense
     val drWho: DrWho
     val dumbAndDumber: DumbAndDumber
     val dune: Dune
@@ -228,6 +229,7 @@ class Faker @JvmOverloads constructor(internal val fakerConfig: FakerConfig = Fa
         dog = Dog(fakerService)
         dota = Dota(fakerService)
         dragonBall = DragonBall(fakerService)
+        drivingLicense = DrivingLicense(fakerService)
         drWho = DrWho(fakerService)
         dumbAndDumber = DumbAndDumber(fakerService)
         dune = Dune(fakerService)