+++ core/src/main/kotlin/io/github/serpro69/kfaker/provider/DrivingLicense.kt
@@ -1,0 +1,17 @@
+package io.github.serpro69.kfaker.provider
+
+import io.github.serpro69.kfaker.*
+import io.github.serpro69.kfaker.dictionary.*
+
+/**
+ * [FakeDataProvider] implementation for [CategoryName.DRIVING_LICENSE] category.
+ */
+@Suppress("unused")
+class DrivingLicense internal constructor(fakerService: FakerService) : AbstractFakeDataProvider<DrivingLicense>(fakerService) {
+    override val categoryName = CategoryName.DRIVING_LICENSE
+    override val localUniqueDataProvider = LocalUniqueDataProvider<DrivingLicense>()
+    override val unique by UniqueProviderDelegate(localUniqueDataProvider)
+
+    fun license() = resolve("usa", "")
+    fun licenseByState(state: String) = resolve("usa", state)
+}