+++ core/src/main/kotlin/io/github/serpro69/kfaker/provider/Rush.kt
@@ -1,0 +1,17 @@
+package io.github.serpro69.kfaker.provider
+
+import io.github.serpro69.kfaker.*
+import io.github.serpro69.kfaker.dictionary.*
+
+/**
+ * [FakeDataProvider] implementation for [CategoryName.RUSH] category.
+ */
+@Suppress("unused")
+class Rush internal constructor(fakerService: FakerService) : AbstractFakeDataProvider<Rush>(fakerService) {
+    override val categoryName = CategoryName.RUSH
+    override val localUniqueDataProvider = LocalUniqueDataProvider<Rush>()
+    override val unique by UniqueProviderDelegate(localUniqueDataProvider)
+
+    fun players() = resolve("players")
+    fun albums() = resolve("albums")
+}