+++ CHANGELOG.md
@@ -24,7 +24,7 @@ and the project versioning adheres to [Semantic Versioning](https://semver.org/s
     - `phish.songs()` - replaces deprecated `song()` function
     - `simpsons.episodeTitles()`
   - Including new providers:
-    - 
+    - `barcode`
 
 ## [v1.4.1] - 2020-08-22
 ### Added