+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and the project versioning adheres to [Semantic Versioning](https://semver.org/s
   - Including new `faker` providers:
     - `barcode`
     - `bigBangTheory`
+    - `drivingLicense`
 
 ## [v1.4.1] - 2020-08-22
 ### Added