+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and the project versioning adheres to [Semantic Versioning](https://semver.org/s
     - `barcode`
     - `bigBangTheory`
     - `drivingLicense`
+    - `drone`
 
 ## [v1.4.1] - 2020-08-22
 ### Added