+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and the project versioning adheres to [Semantic Versioning](https://semver.org/s
     - `futurama`
     - `minecraft`
     - `prince`
+    - `rush`
 
 ## [v1.4.1] - 2020-08-22
 ### Added