+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and the project versioning adheres to [Semantic Versioning](https://semver.org/s
     - `minecraft`
     - `prince`
     - `rush`
+    - `streetFighter`
 
 ## [v1.4.1] - 2020-08-22
 ### Added